pyproject.toml
[project]
name = "adk-copilotkit-quickstart"
version = "0.1.0"
description = "Google ADK Agent CoPilotKit Quickstart Example"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ag-ui-adk>=0.4.2",
    "aiosqlite>=0.20.0",
    "bcrypt>=4.2.0",
    "fastapi>=0.128.4",
    "google-adk>=1.24.1",
    "PyJWT>=2.9.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["adk_web_agent"]
